Move NFF from distribution to xs
