+++ test/celeritas/CMakeLists.txt
@@ -120,8 +120,10 @@ celeritas_add_test(Units.test.cc)
 
 # Distributions
 celeritas_add_test(em/distribution/EnergyLossHelper.test.cc ${_fixme_single})
-celeritas_add_test(em/distribution/NuclearFormFactors.test.cc)
 celeritas_add_test(em/distribution/TsaiUrbanDistribution.test.cc ${_needs_double})
+
+# Cross section calculation
+celeritas_add_test(em/xs/NuclearFormFactors.test.cc)
 
 # Models
 celeritas_add_test(em/BetheHeitler.test.cc ${_needs_double})
