+++ test/celeritas/em/xs/NuclearFormFactors.test.cc
@@ -3,9 +3,9 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/em/distribution/NuclearFormFactors.test.cc
+//! \file celeritas/em/xs/NuclearFormFactors.test.cc
 //---------------------------------------------------------------------------//
-#include "celeritas/em/distribution/NuclearFormFactors.hh"
+#include "celeritas/em/xs/NuclearFormFactors.hh"
 
 #include "celeritas_test.hh"
